+++ examples/visual_fd_demo.rs
@@ -1,12 +1,18 @@
+#![allow(clippy::expect_used)]
+#![allow(clippy::unwrap_used)]
+#![allow(clippy::panic)]
+#![allow(clippy::uninlined_format_args)]
+#![allow(clippy::match_wild_err_arm)]
+#![allow(clippy::cast_ptr_alignment)]
+#![allow(clippy::doc_markdown)]
+
 /// Visual demonstration of file descriptor reuse bug
 ///
 /// This creates output files showing what fd numbers files receive
 /// Run with: cargo run --example visual_fd_demo
+use std::{fs::File, io::Write, os::unix::io::AsRawFd, process::exit};
+
 use fork::{Fork, close_fd, fork, redirect_stdio, waitpid};
-use std::fs::File;
-use std::io::Write;
-use std::os::unix::io::AsRawFd;
-use std::process::exit;
 
 fn main() {
     println!("\nRunning demonstration...\n");